[Refactor] - 테라리움 API PR 반영한 구조 수정
