+++ src/main/java/umc/plantory/domain/terrarium/service/TerrariumCommandUseCase.java
@@ -1,6 +1,6 @@
 package umc.plantory.domain.terrarium.service;
 
-import umc.plantory.domain.terrarium.controller.dto.TerrariumResponseDto;
+import umc.plantory.domain.terrarium.dto.TerrariumResponseDto;
 
 public interface TerrariumCommandUseCase {
     TerrariumResponseDto.WateringTerrariumResponse performTerrariumWatering(String authorization, Long terrariumId);
